f64 -> f32 (#201)

+++ quasim/src/common_test.rs
@@ -1,4 +1,4 @@
-use std::f64::consts::FRAC_1_SQRT_2;
+use std::f32::consts::FRAC_1_SQRT_2;
 
 use nalgebra::{Complex, DVector, dvector};
 
@@ -13,7 +13,7 @@ use crate::{
 
 pub fn double_sub<Sim: Buildable<HybridCircuit> + Debuggable>()
 where
-    Sim::State: QuantumState<BasisValue = Complex<f64>>,
+    Sim::State: QuantumState<BasisValue = Complex<f32>>,
 {
     // Keep for sub circuits
     const N: usize = 2;
@@ -35,7 +35,7 @@ where
         .h(2)
         .h(3);
 
-    const L: usize = 5; //(std::f64::consts::PI * 2f64.sqrt() / 4f64).floor() as usize;
+    const L: usize = 5; //(std::f32::consts::PI * 2f32.sqrt() / 4f32).floor() as usize;
 
     for _ in 0..L {
         circuit = circuit.call("sub", 0);
@@ -107,7 +107,7 @@ where
 
 pub fn deep_sub<Sim: Buildable<HybridCircuit> + Debuggable>()
 where
-    Sim::State: QuantumState<BasisValue = Complex<f64>>,
+    Sim::State: QuantumState<BasisValue = Complex<f32>>,
 {
     // Keep for sub circuits
     const LEVELS: usize = 5;
@@ -136,7 +136,7 @@ where
         forward_steps += 1;
     }
 
-    let mut correct = DVector::<Complex<f64>>::zeros(1 << LEVELS);
+    let mut correct = DVector::<Complex<f32>>::zeros(1 << LEVELS);
     correct[0] = cart!(FRAC_1_SQRT_2);
     correct[1 << (LEVELS - 1)] = cart!(FRAC_1_SQRT_2);
 
@@ -149,7 +149,7 @@ where
         }
         assert_eq!(forward_steps, backward_steps);
 
-        let mut correct = DVector::<Complex<f64>>::zeros(1 << LEVELS);
+        let mut correct = DVector::<Complex<f32>>::zeros(1 << LEVELS);
         correct[0] = cart!(1);
         assert!(equal_state_c(sim.state(), &correct, LEVELS, 0.001));
     }
@@ -157,7 +157,7 @@ where
 
 pub fn hybrid_test<Sim: Buildable<HybridCircuit>>()
 where
-    Sim::State: QuantumState<BasisValue = Complex<f64>>,
+    Sim::State: QuantumState<BasisValue = Complex<f32>>,
 {
     let circuit = Circuit::new(4)
         .new_reg("r0", 1)
@@ -185,7 +185,7 @@ where
     let mut sim = Sim::build(circuit).unwrap();
     sim.run();
 
-    let mut expected = DVector::<Complex<f64>>::zeros(16);
+    let mut expected = DVector::<Complex<f32>>::zeros(16);
     expected[0] = cart!(1.0);
 
     assert!(equal_state_c(sim.state(), &expected, 4, 0.001));
@@ -266,7 +266,7 @@ where
 
 pub fn deep_ctrl_sub<Sim: Buildable<HybridCircuit> + Debuggable>()
 where
-    Sim::State: QuantumState<BasisValue = Complex<f64>>,
+    Sim::State: QuantumState<BasisValue = Complex<f32>>,
 {
     // Keep for sub circuits
     const LEVELS: usize = 5;
@@ -283,7 +283,7 @@ where
         forward_steps += 1;
     }
 
-    let mut correct = DVector::<Complex<f64>>::zeros(1 << LEVELS);
+    let mut correct = DVector::<Complex<f32>>::zeros(1 << LEVELS);
     correct[0b00000] = cart!(FRAC_1_SQRT_2);
     correct[0b00001] = cart!(0.5);
     correct[0b00011] = cart!(-0.35355);
@@ -300,7 +300,7 @@ where
         }
         assert_eq!(forward_steps, backward_steps);
 
-        let mut correct = DVector::<Complex<f64>>::zeros(1 << LEVELS);
+        let mut correct = DVector::<Complex<f32>>::zeros(1 << LEVELS);
         correct[0] = cart!(1);
         assert!(equal_state_c(sim.state(), &correct, LEVELS, 0.001));
     }
